wip: moved menu bar out of listener package [macata #94]
